+++ q_shared.c
@@ -142,10 +142,10 @@ void PerpendicularVector( vec3_t dst, const vec3_t src )
 	*/
 	for ( pos = 0, i = 0; i < 3; i++ )
 	{
-		if ( fabs( src[i] ) < minelem )
+		if ( fabsf( src[i] ) < minelem )
 		{
 			pos = i;
-			minelem = fabs( src[i] );
+			minelem = fabsf( src[i] );
 		}
 	}
 	tempvec[0] = tempvec[1] = tempvec[2] = 0.0F;
@@ -735,7 +735,7 @@ vec_t VectorLength(vec3_t v)
 	
 	length = 0;
 	for (i = 0; i < 3; i++)
-		length += v[i]*v[i];
+		length += v[i] * v[i];
 	length = sqrtf(length);
 
 	return length;
